Added reworked firdt-last benchmarks
