+++ src/it/java/io/deephaven/benchmark/tests/standard/aggby/SortedComboTest.java
@@ -61,21 +61,6 @@ public class SortedComboTest {
     }
 
     @Test
-    public void aggBy4FirstLastOps2Groups() {
-        var aggs = """
-        aggs = [
-           agg.first(['int250']), agg.last(['int640']), 
-           agg.first(['int1M']), agg.last(['str1M'])
-        ]
-        """;
-        runner.addSetupQuery(aggs);
-
-        var q = "source.agg_by(aggs, by=['str250', 'str640'])";
-        runner.test("FirstLast-AggBy- 4 Ops 2 Groups 160K Unique Vals", 160000, q, "str640", "str250", "str1M",
-                "int250", "int640", "int1M");
-    }
-
-    @Test
     public void aggBy4Sorts2Groups() {
         var aggs = """
         aggs = [
